requirements.txt
requests
dnspython
pyftpdlib
aiosmtpd
